cedarpp: tests split, no regressions
